Refactored plotting tests into new PlottingTestHarness class
