+++ tests/testing_harness.py
@@ -2,11 +2,16 @@ from __future__ import print_function
 
 import filecmp
 import hashlib
-from optparse import OptionParser
 import os
 import shutil
 import sys
+import glob
 from collections import OrderedDict
+from optparse import OptionParser
+from PIL import Image
+
+import matplotlib
+import matplotlib.pyplot as plt
 
 sys.path.insert(0, 'openmoc')
 import openmoc
@@ -258,3 +263,53 @@ class TrackingTestHarness(TestHarness):
                      hash_output=False):
         """Return the result string"""
         return self._result
+
+
+class PlottingTestHarness(TestHarness):
+    """Specialized TestHarness for testing plotting."""
+
+    def __init__(self):
+        super(PlottingTestHarness, self).__init__()
+        self.figures = []
+
+    
+    def _get_results(self, num_iters=False, keff=False, fluxes=False,
+                     num_fsrs=False, num_tracks=False, num_segments=False,
+                     hash_output=True):
+
+        outstr = ''
+
+        # Loop over each Matplotlib figure / PIL Image and hash it
+        for i, fig in enumerate(self.figures):
+            plot_filename = 'plot-{0}.png'.format(i)
+
+            # Save the figure to a file
+            if isinstance(fig, matplotlib.figure.Figure):
+                fig.savefig(plot_filename, bbox_inches='tight')
+                plt.close(fig)
+            else:
+                fig.save(plot_filename)
+            
+            # Open the image file in PIL
+            img = Image.open(plot_filename)
+
+            # Hash the image and append to output string
+            plot_hash = hashlib.md5(img.tobytes()).hexdigest()
+            outstr += '{}\n'.format(plot_hash)
+
+        return outstr
+
+    def _cleanup(self):
+        """Delete plot PNG files."""
+
+        # Find all plot files
+        outputs = glob.glob(os.path.join(os.getcwd(), '*.png'))
+
+        # Remove each plot file if it exists
+        for output in outputs:
+            if os.path.isfile(output):
+                os.remove(output)
+            elif os.path.isdir(output):
+                shutil.rmtree(output)
+
+        super(PlottingTestHarness, self)._cleanup()
